Grouping like-with-like, trying to remove duplicate implementations...
