+++ clients/cli/src/orchestrator_client.rs
@@ -1,9 +1,8 @@
 use crate::environment;
-use crate::flops::measure_gflops;
-use crate::memory_stats::get_memory_info;
 use crate::nexus_orchestrator::{
     GetProofTaskRequest, GetProofTaskResponse, NodeType, SubmitProofRequest,
 };
+use crate::utils::system_stats::{get_memory_info, measure_gflops};
 use prost::Message;
 use reqwest::{Client, ClientBuilder};
 use std::time::Duration;
