+++ clients/cli/src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod banner;
+pub mod system_stats;
